reworking the URI structure

+++ application/src/nl/sogeti/android/gpstracker/logger/GPSLoggerServiceManager.java
@@ -99,7 +99,7 @@ public class GPSLoggerServiceManager
       return logging;
    }
 
-   public int startGPSLoggerService(String name)
+   public long startGPSLoggerService(String name)
    {
       connectToGPSLoggerService();
       if ( this.mGPSLoggerRemote != null )
